Fixed problem with Host mode failing with a software timeout when a valid device is re-enumerated. Control request for getting device descriptor was using the device-returned control pipe size, rather than the default size, causing a buffer overflow.
git-svn-id: http://lufa-lib.googlecode.com/svn@151 d5102386-fcda-11dd-9fdb-3debd5008f28

+++ MyUSB/Drivers/USB/LowLevel/HostChapter9.c
@@ -19,7 +19,7 @@ uint8_t USB_Host_SendControlRequest(uint8_t* DataBuffer)
 	uint8_t  ReturnStatus   = HOST_SENDCONTROL_Sucessful;
 	uint8_t* HeaderByte     = (uint8_t*)&USB_HostRequest;
 	uint8_t  DataLen        = USB_HostRequest.DataLength;
-	uint8_t  TimeoutCounter;
+	uint16_t TimeoutCounter;
 
 	USB_HOST_SOFGeneration_Enable();
 	
@@ -44,7 +44,7 @@ uint8_t USB_Host_SendControlRequest(uint8_t* DataBuffer)
 	{
 		if ((ReturnStatus = USB_Host_WaitMS(1)) != HOST_WAITERROR_Sucessful)
 		  goto End_Of_Control_Send;
-	
+
 		if (TimeoutCounter++ == USB_HOST_TIMEOUT_MS)
 		{
 			ReturnStatus = HOST_SENDCONTROL_SoftwareTimeOut;
